% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGGPlot.R
\name{downloadGGPlotButtonServer}
\alias{downloadGGPlotButtonServer}
\title{Download Plot Button (server)}
\usage{
downloadGGPlotButtonServer(id, ggplotObject, height = NULL, width = NULL)
}
\arguments{
\item{id}{Character value that uniquely identifies this instance of the
module. Should match up with the \code{id} value for one
\code{\link[=downloadGGPlotButtonUI]{downloadGGPlotButtonUI()}}.}

\item{ggplotObject}{Reactive expression containing the ggplot object to be
outputted for download.}

\item{height, width}{Height and width of the plot to be outputted}
}
\value{
Downloads the ggplot with the inputted filename
}
\description{
Processes and initiates the download of a ggplot object.
}
