% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTable.R
\name{downloadTableButtonServer}
\alias{downloadTableButtonServer}
\title{Download Plot Button (server)}
\usage{
downloadTableButtonServer(id, dataFrameObject)
}
\arguments{
\item{id}{Character value that uniquely identifies this instance of the
module. Should match up with the \code{id} value for one
\code{\link[=downloadTableButtonUI]{downloadTableButtonUI()}}.}

\item{dataFrameObject}{The name of the table, matrix or data frame object
that is to be exported.}
}
\value{
Downloads the table with the inputted filename
}
\description{
Processes and initiates the download of a table, matrix or data
frame.
}
