% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGGPlot.R
\name{downloadGGPlotButton}
\alias{downloadGGPlotButton}
\title{Download Plot Button (server)}
\usage{
downloadGGPlotButton(
  input,
  output,
  session,
  ggplotObject,
  height = NULL,
  width = NULL
)
}
\arguments{
\item{input, output, session}{Needed for Shiny}

\item{ggplotObject}{Reactive expression containing the ggplot object to be
outputted for download.}

\item{height, width}{Height and width of the plot to be outputted}
}
\value{
Downloads the ggplot with the inputted filename
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use
\code{\link[=downloadGGPlotButtonServer]{downloadGGPlotButtonServer()}} instead.

Processes and initiates the download of a ggplot object.
}
